+++ MVCDemo/MVCDemo/Models/Person.cs
@@ -12,7 +12,7 @@ namespace MVCDemo {
         protected String _FirstName = "";
         protected String _LastName = "";
         protected DateTime _BirthDate = DateTime.MaxValue;
-        protected Double _Height = 0.0;
+        protected decimal _Height = 0.0M;
         protected float _ShoeSize = 0.0F;
         protected Color _EyeColor = Color.Brown;
         // 0 = brown, 1 = blue, 
@@ -132,7 +132,7 @@ namespace MVCDemo {
         }
 
 
-        public virtual double Height {
+        public virtual decimal Height {
             get {
                 return _Height;
             }
