+++ MVCDemo/MVCDemo/Models/Villian.cs
@@ -154,7 +154,7 @@ namespace MVCDemo {
             _LastName = (string)dr[db_LastName];
             _BirthDate = (DateTime)dr[db_DateOfBirth];
             _EyeColor = (Color)Enum.ToObject(typeof(Color), (byte)dr[db_EyeColor]);
-            _Height = (double)dr[db_HeightInInches];
+            _Height = (decimal)dr[db_HeightInInches];
             _SideKickID = (int)dr[db_SideKick];
             _CostumeID = (int)dr[Costume.db_ID];
         }
